tests/turbohtml-tests/treebuilder_coverage.dat

#data
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
#errors
#document
| <!DOCTYPE html "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|   <body>

#data
\x00\x00\x00
#errors
#document
| <html>
|   <head>
|   <body>

#data
<head></head>\x00text
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<head></head>\x0ctext
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<math><mi>\x00text</mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "text"

#data
<math><mi>\x0ctext</mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "text"

#data
  text
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<table>\x00\x00</table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>

#data
<head></head></template>text
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<title></title>
#errors
#document
| <html>
|   <head>
|     <title>
|   <body>

#data
<dl><dd>x</dd></dl>
#errors
#document
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "x"

#data
<select><optgroup><optgroup></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|       <optgroup>

#data
<body>\x00</body>
#errors
#document
| <html>
|   <head>
|   <body>

#data
<a><table><a>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>

#data
<nobr><nobr>x</nobr></nobr>
#errors
#document
| <html>
|   <head>
|   <body>
|     <nobr>
|     <nobr>
|       "x"


#data
<select><option><optgroup></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <optgroup>

#data
<table><caption><!--comment--></caption></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <!-- comment -->

#data
<table><colgroup></colgroup><div>x</div></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|     <table>
|       <colgroup>

#data
<table><colgroup></col></colgroup></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>

#data
<table><tbody><tr></tbody></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<table><tr></tr></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<table><tr><td></tr></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<select><option></option></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<select><optgroup></optgroup></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>

#data
<select></optgroup></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>


#data
<table><caption></tbody></caption></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
<table><caption></table></caption>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
<table><tr><td></tbody></td></tr></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<select>\x00text</select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       "text"

#data
<select>\x0ctext</select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       "text"

#data
<select><!--comment--></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <!-- comment -->

#data
<select><option><select></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<template><td></template>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <td>
|   <body>

#data
<template><caption></template>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <caption>
|   <body>

#data
<template><colgroup></template>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <colgroup>
|   <body>


#data
<frameset><frameset></frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>
|     <frameset>

#data
<frameset></frameset><noframes>x</noframes>
#errors
#document
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "x"

#data
</body></html><!--after-->
#errors
#document
| <html>
|   <head>
|   <body>
| <!-- after -->

#data
</body></html><div>x</div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"

#data
<table><tbody></tr></tbody></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>

#data
<table><tr></td></tr></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<table><tr><td></th></td></tr></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<html class="x"><body class="y">
#errors
#document
| <html>
|   class="x"
|   <head>
|   <body>
|     class="y"

#data
<table>   </table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       "   "

#data
text
#errors
#document-fragment
td
#document
| "text"

#data
<tr><td>x</td></tr>
#errors
#document-fragment
tbody
#document
| <tr>
|   <td>
|     "x"

#data
<td>x</td>
#errors
#document-fragment
tr
#document
| <td>
|   "x"

#data
<option>x</option>
#errors
#document-fragment
select
#document
| <option>
|   "x"

#data
<col>
#errors
#document-fragment
colgroup
#document
| <col>

#data
text
#errors
#document-fragment
table
#document
| "text"

#data
<svg><path></path></svg>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg path>


#data
<table>\x00text</table>
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"
|     <table>

#data
<table>\x0ctext</table>
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"
|     <table>

#data
<frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>

#data
<svg><!DOCTYPE html></svg>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
</frameset>
#errors
#document-fragment
frameset
#document

#data
<frameset></frameset></frameset>
#errors
#document
| <html>
|   <head>
|   <frameset>


#data
<table>\x00</table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>

#data
<tr>
#errors
#document-fragment
caption
#document

#data
<table>
#errors
#document-fragment
caption
#document
| <table>

#data
<table><caption><table></table></caption></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|     <table>

#data
<table>\x00</table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
